# Speed up text generation, remove faker from text config builder