Caching of text configs brings about 20% of speed
before
```
Benchmark                                         Mode  Cnt     Score    Error   Units
DatafakerPassword.password10                     thrpt   10  1293.016 ± 38.115  ops/ms
DatafakerPassword.password10WithUpper            thrpt   10  1059.780 ± 83.480  ops/ms
DatafakerPassword.password10WithUpperLower       thrpt   10   963.004 ± 79.944  ops/ms
DatafakerPassword.password10WithUpperLowerDigit  thrpt   10   992.366 ± 32.792  ops/ms
```
after
```

Benchmark                                         Mode  Cnt     Score     Error   Units
DatafakerPassword.password10                     thrpt   10  1577.841 ± 112.689  ops/ms
DatafakerPassword.password10WithUpper            thrpt   10  1262.548 ±  74.436  ops/ms
DatafakerPassword.password10WithUpperLower       thrpt   10  1212.872 ±  27.308  ops/ms
DatafakerPassword.password10WithUpperLowerDigit  thrpt   10  1203.032 ±  45.527  ops/ms
```